Continuation of core API refactoring and simplification.
--HG--
rename : praxis.core/src/net/neilcsmith/praxis/core/PortProxyContainer.java => praxis.core/src/net/neilcsmith/praxis/core/ContainerContext.java
rename : praxis.core/src/net/neilcsmith/praxis/core/ProxyRegistrationException.java => praxis.core/src/net/neilcsmith/praxis/core/RegistrationException.java
rename : praxis.core/src/net/neilcsmith/praxis/core/ServiceManager.java => praxis.core/src/net/neilcsmith/praxis/core/interfaces/ServiceManager.java
rename : praxis.core/src/net/neilcsmith/praxis/core/ServiceUnavailableException.java => praxis.core/src/net/neilcsmith/praxis/core/interfaces/ServiceUnavailableException.java

+++ rapl/src/net/neilcsmith/rapl/render/BusClient.java
@@ -171,7 +171,7 @@ public class BusClient implements Bus, AudioClient {
             float[] data = sinks[i].buffer.getData();
             FloatBuffer out = outputs.get(i);
             for (int f=0; f<nframes; f++) {
-                out.put(f, (float) data[f]);
+                out.put(f, data[f]);
             }
         }
     }
